DAT-4340 add no scoll mixin for body state handling
